Update infra plan workflow to fix incomplete PR comments (#605)
This PR fixes the `infra_plan` workflow issue where the PR comment was
too large and got truncated.
If the plan output exceeds `60KB`, only the final summary of the plan
will be printed in the comment, along with a message indicating to check
the action logs for the full output:

![image](https://github.com/user-attachments/assets/7cc560c8-5431-4a37-8cec-e9117fd5d9c7)
Otherwise, it works as before, but with a cleaner output as all
unnecessary lines have been removed.
In addition fix the possible code injection founded by CodeQL.

Resolves: CES-1061