Pin `azure-function-core-tools` version (#629)
TL;DR: Pin `azure-functions-core-tools` to version `4.0.7030`

[Microsoft is currently working on an official Linux ARM64 release (now
in
**Preview**)](https://github.com/Azure/azure-functions-core-tools/releases/tag/4.0.7332-preview1),
which will eventually be available via **NPM** or **APT**, eliminating
the need to manually build the CLI from source.

To support this platform, recent changes were made to the source code
that are incompatible with the build steps used by this **devcontainer
feature**.

As a result, we will now only support version `4.0.7030`, which is the
latest version compatible with our `install.sh` script.

Once the official build is released, this feature will be **deprecated**
and replaced by the [community-maintained
version](https://github.com/jlaundry/devcontainer-features/tree/main/src/azure-functions-core-tools).

---

I decided to mark this change as a patch to avoid breaking the
devcontainer builds used by our teams — since the `version` parameter
(which I removed) was optional and not used by anyone. Releasing this
change as v2 would have inevitably broken local builds.