Update PEP creation logic for Postgres module (#764)
This change updates the logic for creating the private endpoint in the
Azure PostgreSQL server module. It introduces a new variable
`delegated_subnet_id` to allow users to specify either `subnet_pep_id`
or `delegated_subnet_id`, but not both. The private endpoint output is
now optional and will return null if the private endpoint is not
created. This change maintains backward compatibility, but the old
output structure will be removed in the next major version.

Resolves: CES-1165

---------

Co-authored-by: Christian Calabrese <christian.calabrese@pagopa.it>