Add SBOMs and generation scripts (#804)
This PR introduces:

- Script to `generate` and `validate` SBOMs using
[syft](https://github.com/anchore/syft) and
[grype](https://github.com/anchore/grype)
- Integration of `syft` and `grype` into the `devcontainer`
- A new `sboms/` folder containing all generated **SBOM** files
- Only manual run for SBOM generation and validation

Each SBOM is generated per module, provider, or workspace to allow finer
granularity, avoid duplication, and simplify future validation or
comparison.

An alternative implementation using a single unified SBOM file is
available in the
[feat-create-sbom-file-alternative](https://github.com/pagopa/dx/tree/feat-create-sbom-file-alternative)
branch, but it may miss some components and lacks the same level of
control.

Closes: CES-1205
Relates: CES-1040