Remove hardcoded version of pre-commit from template (#901)
This pull request updates the monorepo generator so that the revision
for the `pre-commit-terraform` hook is fetched dynamically from GitHub,
ensuring that generated repositories always use the latest available
version. The change introduces a new action to retrieve the latest
release, exposes it to Handlebars templates, and updates the template to
use the dynamic version.

Closes CES-1276