Fix Static Web App workflow blocked deploy by branch protection (#873)
This PR updates the GitHub Actions workflow to fix the issue where
preview deployments were not triggered because the deploy job was
running on an environment protected by branch rules on main.
Now, the deploy job runs on CI environment when PR is created and
updated, otherwise (when is merged) it run on CD environment.
A new step has been added to remove the preview when the PR is closed.

Resolves: CES-1271
depends-on: #905 CES-1270