Remove `no-console` rule (#920)
As it is stated in the official
documentation (https://eslint.org/docs/latest/rules/no-console),
`no-console` is a rule that is meant for code that targets to
`browsers`.

Our `@pagopa/eslint-config` package is a generic, base package for js
projects so it should contain only the common rules shared by all
environments.