Configure DevContainer (#923)
This pull request introduces DevContainer support to the monorepo
generator, making it easier to set up a consistent development
environment for new and existing team members. The main changes include
installing the `@devcontainers/cli` dependency, adding a new action to
configure DevContainers, and updating documentation to guide users on
using DevContainers with Visual Studio Code or the console.

Closes CES-1255