Fix types export (#940)
This pull request updates the `package.json` configuration for the
`packages/azure-tracing` package to improve TypeScript support and
compatibility with different module systems. The main changes involve
adding type definitions and CommonJS entry points to the package
exports, as well as introducing a `typesVersions` field for more precise
type resolution.

Closes CES-1307 #921