Fallback to latest when RUNNER_DIGEST is not provided (#983)
`docker build` for `self-hosted-runner` fails with
"InvalidDefaultArgInFrom" because when `RUNNER_DIGEST` is not provided (via `--build-arg`) the
resulting `FROM` content is `FROM ghcr.io/actions/actions-runner@ as
base`, which is invalid.

With this change, I added `@` as prefix only when `RUNNER_DIGEST` is
provided so `docker build .` without `--build-arg` will result to `FROM
ghcr.io/actions/actions-runner AS base` that defaults to `latest` tag.