Add a random revision suffix to containerapp (#1026)
`az containerapp revision copy`, by default, copies the given revision
by creating a new one using the same name with an incremented numerical
suffix (`my-revision-001` will be copied into `my-revision-002`).

Since our deployment mechanism always copies the last active revision,
when a rollback occurs the subsequent deploy fails because it tries to
create a revision with a name that is already taken.

With this change we override the suffix to use a random (4-char) string.

Closes CES-1375