Update docker build/push action with tag (#1042)
This pull request enhances the Docker build and push GitHub Action
workflow by improving how version tags are extracted and passed to the
Docker metadata step. The main goal is to ensure that Docker images are
tagged more accurately based on the tag that triggered the workflow,
supporting both `v1.2.3` and `1.2.3` formats.