Add `onBrokenAnchors` option (#1047)
This pull request introduces a configuration update to the documentation
website to improve link validation during the build process. The main
change ensures that builds will fail if there are any broken anchor
links, helping maintain the quality and reliability of site navigation.

With this option, we ensure that the deployed website does not contain
any invalid links.

Depends on #1048