Add WAF to MCP server api gateway (#1127)
The WAF cannot be directly attached to an HTTP API Gateway. The
alternatives were switching to a classic REST API Gateway or placing a
CloudFront distribution in front of the existing HTTP API Gateway.
After evaluating the limitations of securing the native API Gateway
domain behind CloudFront, the REST API Gateway has been chosen as the
safer and more reliable option.
Already applied for testing

Resolves CES-1475