Update `azure_storage_account` with infrastructure encryption override (#1154)
This pull request introduces a **patch** to the Azure storage account
module to address a breaking change caused by the default enabling of
**infrastructure encryption** for the `audit` use case.
The main update is the addition of a temporary
`override_infrastructure_encryption` variable, which allows users to
disable infrastructure encryption and prevent forced recreation of
existing storage accounts.
This variable is deprecated and will be removed in the next **major**
version.
The changes also ensure backward compatibility and update documentation,
validation logic, and tests accordingly.

Resolves: CES-1540